@@ -1,158 +1,114 @@
 [build-system]
-requires = ["setuptools>=61.2", "setuptools_scm[toml]>=7.0"]
 build-backend = "setuptools.build_meta"
+requires = [
+    "setuptools>=61.2",
+    "setuptools_scm[toml]>=8.0",
+]
 
 [project]
 name = "tmmc-lnpy"
-authors = [{ name = "William P. Krekelberg", email = "wpk@nist.gov" }]
-license = { text = "NIST-PD" }
 description = "Analysis of lnPi results from TMMC simulation"
-# if using markdown
-# long_description_content_type = text/markdown
-keywords = ["tmmc-lnpy"]
+keywords = [
+    "tmmc-lnpy",
+]
+license = { text = "NIST-PD" }
+authors = [{ name = "William P. Krekelberg", email = "wpk@nist.gov" }]
+requires-python = ">=3.8,<=3.12"
 classifiers = [
     "Development Status :: 2 - Pre-Alpha",
+    "Intended Audience :: Science/Research",
     "License :: Public Domain",
     "Operating System :: OS Independent",
-    "Intended Audience :: Science/Research",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3 :: Only",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Scientific/Engineering",
 ]
-dynamic = ["readme", "version"]
-requires-python = ">=3.8"
+dynamic = [
+    "readme",
+    "version",
+]
 dependencies = [
-    "numpy",
-    "xarray",
-    "scipy",
-    "scikit-image",
-    "joblib",
     "bottleneck",
-    "tqdm",
     "importlib_resources",
-    "module-utilities >= 0.9.0",
+    "joblib",
     "lazy_loader",
-] # additional packages
-
-[project.urls]
-Homepage = "https://github.com/usnistgov/tmmc-lnpy"
-Documentation = "https://pages.nist.gov/tmmc-lnpy/"
-
+    "module-utilities >= 0.9.0",
+    "numpy",
+    "scikit-image >= 0.21",
+    "scipy",
+    "tqdm",
+    "xarray",
+]
 [project.optional-dependencies]
-test = [
-    "pytest", #
-    "pytest-xdist",
-    "pytest-cov",
-    "pytest-sugar",
-]
-viz = [
-    "matplotlib-base", #
-    "ipywidgets",
-]
-dev-extras = [
-    "setuptools-scm", #
-    "pytest-accept", # p2c: -p
-    "ipython",
-    "ipykernel",
-    "tmmc-lnpy[viz]",
-    "nbval",
-]
-typing-extras = [
-    "pytype; python_version < '3.11'", #
-    "mypy >= 1.4.1",
-    "pandas-stubs",
-    "types-tqdm",
-]
-typing = [
-    "tmmc-lnpy[typing-extras]", #
-    "tmmc-lnpy[test]",
-]
-nox = [
-    "nox",
-    "noxopt", # p2c: -p
-    "ruamel.yaml",
-]
 dev = [
-    "tmmc-lnpy[test]", #
+    "tmmc-lnpy[nbval]",
+    "tmmc-lnpy[test]",
     "tmmc-lnpy[typing-extras]",
     "tmmc-lnpy[dev-extras]",
-]
-tools = [
-    "pre-commit", #
-    # "cruft",
-    "scriv", # p2c: -p
-    "nbqa",
-    "pyright",
-]
-dev-complete = [
-    "tmmc-lnpy[dev]", #
-    "tmmc-lnpy[tools]",
-    "tmmc-lnpy[nox]",
-]
-docs = [
-    "ipython", #
-    "pyenchant",
-    "ghp-import",
-    "sphinx >= 5.3.0",
-    "sphinx-copybutton",
-    "sphinxcontrib-spelling",
-    "sphinx-autobuild",
-    "myst-nb",
-    "sphinx-book-theme",
-    "autodocsumm",
     "tmmc-lnpy[viz]",
 ]
-# to be parsed with pyproject2conda with --no-base option
-dist-pypi = ["twine", "build"]
-dist-conda = [
-    "anaconda-client", #
-    "grayskull",
-    "conda-build",
-    "conda-verify",
-    "boa",
+dev-complete = [
+    "tmmc-lnpy[dev]",
+    "tmmc-lnpy[tools]",
 ]
-
-[tool.pyproject2conda]
-user_config = "config/userconfig.toml"
-template_python = "requirements/py{py}-{env}"
-template = "requirements/{env}"
-# defaults
-python = ["3.10"]
-channels = ["conda-forge"]
-style = ["yaml", "requirements"]
-default_envs = [
-    "test",
-    "typing",
-    "docs",
-    "dev",
-    "dev-complete",
-    "dist-conda",
-    "dist-pypi"
+dev-extras = [
+    "ipykernel",
+    "ipython",
+    "nox",
+    "pytest-accept",
 ]
+docs = [
+    "autodocsumm",
+    "ghp-import",
+    "ipython",
+    "myst-nb",
+    "pyenchant",
+    "sphinx >= 5.3.0",
+    "sphinx-autobuild",
+    "sphinx-book-theme",
+    "sphinx-copybutton",
+    "sphinxcontrib-spelling",
+    "tmmc-lnpy[viz]",
+]
+nbval = [
+    "nbval",
+]
+test = [
+    "pytest",
+    "pytest-cov",
+    "pytest-sugar",
+    "pytest-xdist",
+]
+tools = [
+    "nbqa",
+    "pip-tools",
+    "pre-commit",
+    "pyright",
+    "scriv",
+]
+typing = [
+    "pytest",
+    "tmmc-lnpy[typing-extras]",
+]
+typing-extras = [
+    "mypy >= 1.6.1",
+    "pandas-stubs",
+    "pytype; python_version < '3.11'",
+    "types-tqdm",
+]
+viz = [
+    "ipywidgets",
+    "matplotlib",
+]
+[project.urls]
+Documentation = "https://pages.nist.gov/tmmc-lnpy/"
+Homepage = "https://github.com/usnistgov/tmmc-lnpy"
 
-[tool.pyproject2conda.envs.test-extras]
-extras = ["test"]
-
-[tool.pyproject2conda.envs.dev-base]
-extras = ["dev"]
-
-[[tool.pyproject2conda.overrides]]
-envs = ["test-extras", "dist-conda", "dist-pypi"]
-base = false
-
-[[tool.pyproject2conda.overrides]]
-envs = ["test", "typing", "test-extras"]
-python = ["3.8", "3.9", "3.10", "3.11"]
-
-[[tool.pyproject2conda.overrides]]
-envs = ["dist-conda"]
-style = ["yaml"]
-
-## grayskull still messes some things up, but use scripts/recipe-append.sh for this
 [tool.setuptools]
 zip-safe = true # if using mypy, must be False
 include-package-data = true
@@ -396,3 +351,57 @@ reportUnboundVariable = false
 
 [tool.pytype]
 inputs = ["src", "tests"]
+
+# * Dependency creation --------------------------------------------------------
+
+[tool.pyproject2conda]
+user_config = "config/userconfig.toml"
+template_python = "requirements/py{py}-{env}"
+template = "requirements/{env}"
+python = ["3.11"]
+channels = ["conda-forge"]
+style = ["yaml", "requirements"]
+default_envs = ["test", "typing", "docs", "dev", "dev-complete"]
+
+[tool.pyproject2conda.dependencies]
+pytest-accept = { pip = true }
+scriv = { pip = true }
+matplotlib = { skip = true, packages = "matplotlib-base" }
+
+[tool.pyproject2conda.envs.build]
+extras = "build-system.requires"
+reqs = ["build"]
+base = false
+style = "requirements"
+
+[tool.pyproject2conda.envs.test-extras]
+extras = ["test"]
+base = false
+
+[tool.pyproject2conda.envs.test-notebook]
+extras = ["test", "nbval", "viz"]
+
+[[tool.pyproject2conda.overrides]]
+# Ensure these have pip in yaml file
+envs = ["test", "docs", "dev", "dev-complete"]
+deps = ["pip"]
+
+[[tool.pyproject2conda.overrides]]
+envs = ["test", "typing", "test-extras"]
+python = ["3.8", "3.9", "3.10", "3.11", "3.12"]
+
+[tool.pip-tools]
+allow-unsafe = true
+strip-extras = true
+
+# * Other tools ----------------------------------------------------------------
+
+[tool.scriv]
+format = "md"
+md_header_level = "2"
+new_fragment_template = "file: changelog.d/templates/new_fragment.md.j2"
+
+[tool.commitizen]
+use_shortcuts = true
+
+[tool.cruft]
