@@ -31,7 +31,7 @@ if TYPE_CHECKING:
 
     _FillArg = Union[int, None]
     _FillVal = Union[_FillArg, float]
-    _ExtremaArg = Union[int, "tuple[int, ...]", None]  # noqa
+    _ExtremaArg = Union[int, "tuple[int, ...]", None]
 
 
 @docfiller.decorate
@@ -112,7 +112,7 @@ def find_boundaries_overlap(
 
 
 @docfiller.decorate
-def find_boundaries_overlap(
+def find_boundaries_overlap(  # noqa: C901
     masks: Sequence[MyNDArray],
     *,
     boundaries: list[MyNDArray] | None = None,
@@ -152,7 +152,7 @@ def find_boundaries_overlap(
     """
 
     n = len(masks)
-    assert method in ["approx", "exact"]
+    assert method in {"approx", "exact"}
 
     if boundaries is None:
         boundaries = find_boundaries(masks, mode=mode, connectivity=connectivity)
@@ -184,10 +184,11 @@ def find_boundaries_overlap(
 
     if method == "approx":
         return _get_approx()
-    elif method == "exact":
+    if method == "exact":
         return _get_exact()
-    else:
-        raise ValueError(f"unknown method={method}")
+
+    msg = f"unknown method={method}"
+    raise ValueError(msg)
 
 
 @docfiller.decorate
@@ -234,7 +235,8 @@ def find_masked_extrema(
     elif extrema == "min":
         func = np.argmin
     else:
-        raise ValueError('extrema must be on of {"min", "max}')
+        msg = 'extrema must be on of {"min", "max}'
+        raise ValueError(msg)
 
     masks = masks_change_convention(masks, convention, "image")
 
@@ -254,10 +256,10 @@ def find_masked_extrema(
         else:
             mask_flat = mask.reshape(-1)
             arg = positions_flat[mask_flat][func(data_flat[mask_flat])]
-            val = data_flat[arg]  # type: ignore
+            val = data_flat[arg]  # type: ignore[assignment]
 
             if unravel:
-                arg = np.unravel_index(arg, data.shape)  # type: ignore
+                arg = np.unravel_index(arg, data.shape)  # type: ignore[assignment,arg-type]
 
         out_arg.append(arg)
         out_val.append(val)
@@ -323,8 +325,8 @@ def merge_regions(
     w_min = np.array(w_min, copy=True)
 
     # keep track of keep/kill
-    mapping = {i: msk for i, msk in enumerate(masks)}
-    for cnt in range(nfeature):
+    mapping = dict(enumerate(masks))
+    for _cnt in range(nfeature):
         # number of finite minima
         nfeature = len(mapping)
 
@@ -342,7 +344,7 @@ def merge_regions(
                         stacklevel=2,
                     )
                 break
-            elif nfeature <= nfeature_max:
+            if nfeature <= nfeature_max:
                 break
 
         idx_keep, idx_kill = min_arg
@@ -361,7 +363,7 @@ def merge_regions(
 
         # new mask
         mapping[idx_keep] |= mapping[idx_kill]  # type: ignore[index]
-        del mapping[idx_kill]  # type: ignore
+        del mapping[idx_kill]  # type: ignore[arg-type]
 
     # from mapping create some new stuff
     # new w/de
@@ -381,7 +383,7 @@ def merge_regions(
 
 
 @docfiller.decorate
-class wFreeEnergy:
+class wFreeEnergy:  # noqa: N801
     r"""
     Analysis of local free energy :math:`w = \beta f = - \ln \Pi`.
 
@@ -521,15 +523,15 @@ class wFreeEnergy:
             valmax_dict = dict(zip(overlap_keys, valmax))
 
             # loop over unique keys
-            for i, j in {(i, j) for i, j, _ in overlap}:  # type: ignore
+            for i, j in {(i, j) for i, j, _ in overlap}:  # type: ignore[misc]
                 vals = [valmax_dict[i, j, index] for index in range(2)]
                 # take min value of maxes
                 if np.all(np.isnan(vals)):
                     out_arg[i, j] = None
                 else:
                     idx_min = np.nanargmin(vals)
-                    out_arg[i, j] = argmax_dict[i, j, idx_min]  # type: ignore
-                    out_max[i, j] = out_max[j, i] = valmax_dict[i, j, idx_min]  # type: ignore
+                    out_arg[i, j] = argmax_dict[i, j, idx_min]  # type: ignore[index]
+                    out_max[i, j] = out_max[j, i] = valmax_dict[i, j, idx_min]  # type: ignore[index]
         return out_arg, out_max
 
     @property
@@ -615,13 +617,13 @@ def _get_w_data(index: pd.MultiIndex, w: wFreeEnergy) -> dict[str, pd.Series[Any
     w_argmin = pd.Series(w.w_argmin, index=w_min.index, name="w_argmin")
 
     w_tran = (
-        pd.DataFrame(
+        pd.DataFrame(  # type: ignore[call-overload]
             w.w_tran,
             index=index,
             columns=index.get_level_values("phase").rename("phase_nebr"),
         )
         .stack()
-        .rename("w_tran")  # type: ignore
+        .rename("w_tran")
     )
 
     # get argtrans values for each index
@@ -652,7 +654,7 @@ def _get_w_data(index: pd.MultiIndex, w: wFreeEnergy) -> dict[str, pd.Series[Any
     }  # [index_map, w.w_argtran]}
 
 
-class wFreeEnergyCollection:
+class wFreeEnergyCollection:  # noqa: N801
     r"""
     Calculate the transition free energies for a :class:`lnpy.lnpiseries.lnPiCollection`.
 
@@ -693,9 +695,7 @@ class wFreeEnergyCollection:
             _get_w_data, items=seq, use_joblib=self._use_joblib, total=len(ws)
         )
 
-        result = {key: pd.concat([x[key] for x in out]) for key in out[0].keys()}
-
-        return result
+        return {key: pd.concat([x[key] for x in out]) for key in out[0]}
 
     @property
     def w_min(self) -> pd.Series[Any]:
@@ -766,8 +766,7 @@ class wFreeEnergyCollection:
             nebrs = delta_w.indexes["phase_nebr"].intersection(idx_nebr)
             delta_w = delta_w.sel(phase=idx, phase_nebr=nebrs)
 
-        out = delta_w.min("phase_nebr").fillna(0.0)
-        return out
+        return delta_w.min("phase_nebr").fillna(0.0)
 
     def get_dw(
         self, idx: int, idx_nebr: int | list[int] | None = None
@@ -777,7 +776,7 @@ class wFreeEnergyCollection:
 
 
 # @lnPiCollection.decorate_accessor("wfe_phases")
-class wFreeEnergyPhases(wFreeEnergyCollection):
+class wFreeEnergyPhases(wFreeEnergyCollection):  # noqa: N801
     """
     Stripped down version of :class:`wFreeEnergyCollection` for single phase grouping.
 
@@ -811,7 +810,9 @@ class wFreeEnergyPhases(wFreeEnergyCollection):
         """Series representation of delta_w"""
         return self.dwx.to_series()
 
-    def get_dw(self, idx: int, idx_nebr: int | Iterable[int] | None = None) -> float | MyNDArray:  # type: ignore[override]
+    def get_dw(  # type: ignore[override]
+        self, idx: int, idx_nebr: int | Iterable[int] | None = None
+    ) -> float | MyNDArray:
         dw = self.dwx
         index = dw.indexes["phase"]
 
