@@ -1,3 +1,4 @@
+# pyright: reportPrivateUsage=false
 """
 Ensemble averages (:mod:`~lnpy.ensembles`)
 ==========================================
@@ -45,7 +46,7 @@ def _get_range(n: int) -> MyNDArray:
 
 
 @lru_cache(maxsize=10)
-def get_xrlnPiWrapper(
+def get_xrlnpiwrapper(
     shape: tuple[int, ...],
     rec_name: str | None = "sample",
     n_name: str = "n",
@@ -63,7 +64,7 @@ def get_xrlnPiWrapper(
     )
 
 
-class xlnPiWrapper:
+class xlnPiWrapper:  # noqa: N801
     """
     Wraps lnPi objects with xarray functionality
 
@@ -79,7 +80,7 @@ class xlnPiWrapper:
         n_name: str = "n",
         lnz_name: str = "lnz",
         comp_name: str = "component",
-        phase_name: str = "phase",
+        phase_name: str = "phase",  # noqa: ARG002
     ) -> None:
         self.shape = shape
         self.ndim = len(shape)
@@ -113,16 +114,13 @@ class xlnPiWrapper:
             "dims_state": self.dims_lnz + list(args),
         }
 
-        if self.dims_rec is not None:
+        if self.dims_rec:
             d["dims_rec"] = self.dims_rec
         return d
 
     @cached_meth
     def ncoords(self, coords_n: bool = False) -> xr.DataArray:
-        if coords_n:
-            coords = self.coords_n
-        else:
-            coords = None
+        coords = self.coords_n if coords_n else None
         return xr.DataArray(
             _get_indices(self.shape),
             dims=self.dims_comp + self.dims_n,
@@ -182,23 +180,16 @@ def xr_name(
     """Decorator to add name, longname to xarray output"""
 
     def decorator(
-        func: C_Ensemble[T_Ensemble, P, xr.DataArray]
+        func: C_Ensemble[T_Ensemble, P, xr.DataArray],
     ) -> C_Ensemble[T_Ensemble, P, xr.DataArray]:
-        if name is None:
-            _name = func.__name__.lstrip("_")
-        else:
-            _name = name
+        _name = func.__name__.lstrip("_") if name is None else name
 
         @wraps(func)
         def wrapper(
             self: T_Ensemble, /, *args: P.args, **kwargs: P.kwargs
         ) -> xr.DataArray:
             out = func(self, *args, **kwargs).rename(_name)
-            if hasattr(self, "_standard_attrs"):
-                attrs = self._standard_attrs
-            else:
-                attrs = {}
-            attrs = dict(attrs, **kws)
+            attrs = dict(getattr(self, "_standard_attrs", {}), **kws)
 
             if long_name is not None:
                 attrs["long_name"] = long_name
@@ -218,7 +209,7 @@ def xr_name(
 #     return xGrandCanonical(parent)
 
 
-class xGrandCanonical:
+class xGrandCanonical:  # noqa: PLR0904,N801
     """
     :class:`~xarray.DataArray` accessor to Grand Canonical properties from lnPi
 
@@ -238,7 +229,7 @@ class xGrandCanonical:
             self._rec_name = None
             first = parent
 
-        self._wrapper = get_xrlnPiWrapper(shape=first.shape, rec_name=self._rec_name)
+        self._wrapper = get_xrlnpiwrapper(shape=first.shape, rec_name=self._rec_name)
         self._cache: dict[str, Any] = {}
 
     @property
@@ -257,19 +248,19 @@ class xGrandCanonical:
     def first(self) -> lnPiMasked:
         if isinstance(self._parent, lnPiCollection):
             return self._parent.iloc[0]
-        else:
-            return self._parent
+
+        return self._parent
 
     # @cached_prop to much memory
     @property
     def _rec_coords(self) -> dict[str, pd.Index[Any] | pd.MultiIndex | float | int]:
         if isinstance(self._parent, lnPiMasked):
             return dict(self._parent._index_dict(), **self._parent.state_kws)
-        else:
-            return {
-                self._parent._concat_dim: self._parent.index,
-                **self._parent.state_kws,
-            }
+
+        return {
+            self._parent._concat_dim: self._parent.index,
+            **self._parent.state_kws,
+        }
 
     @property
     def _xarray_dot_kws(self) -> dict[str, str]:
@@ -303,7 +294,7 @@ class xGrandCanonical:
     @property
     def dims_state(self) -> list[str]:
         """Dimensions corresponding to 'state variables'"""
-        return self.pi_norm.attrs["dims_state"]  # type: ignore
+        return self.pi_norm.attrs["dims_state"]  # type: ignore[no-any-return]
 
     @property
     def dims_rec(self) -> list[str]:
@@ -313,12 +304,12 @@ class xGrandCanonical:
     @property
     def beta(self) -> float:
         r"""Inverse temperature :math:`\beta = 1 / (k_{\rm B} T)`"""
-        return self._parent.state_kws["beta"]  # type: ignore
+        return self._parent.state_kws["beta"]  # type: ignore[no-any-return]
 
     @property
     def volume(self) -> float:
         """System volume :math:`V`."""
-        return self._parent.state_kws["volume"]  # type: ignore
+        return self._parent.state_kws["volume"]  # type: ignore[no-any-return]
 
     @cached_prop
     def coords_state(self) -> dict[str, xr.DataArray]:
@@ -384,7 +375,7 @@ class xGrandCanonical:
     def lnpi_norm(self) -> xr.DataArray:
         r""":math:`\ln \Pi_{\rm norm}(N)`."""
         pi = self.pi_norm
-        return np.log(xr.where(pi > 0, pi, np.nan))  # type: ignore
+        return np.log(xr.where(pi > 0, pi, np.nan))  # type: ignore[no-untyped-call,no-any-return]
 
     def _get_prop_from_extra_kws(
         self, prop: str | ArrayLike | xr.DataArray
@@ -393,9 +384,10 @@ class xGrandCanonical:
             p = self.first
 
             if prop not in p.extra_kws:
-                raise ValueError(f"{prop} not found in extra_kws")
-            else:
-                prop = p.extra_kws[prop]
+                msg = f"{prop} not found in extra_kws"
+                raise ValueError(msg)
+
+            prop = p.extra_kws[prop]
 
         # if not isinstance(prop, (np.ndarray, xr.DataArray)):
         #     raise ValueError(
@@ -422,8 +414,8 @@ class xGrandCanonical:
             x = xr.DataArray(x, dims=self.dims_n)
         return x
 
-    def _mean_pi(self, x: xr.DataArray, **kwargs: Any) -> xr.DataArray:
-        return xr.dot(self.pi_norm, x, dims=self.dims_n, **self._xarray_dot_kws)  # type: ignore
+    def _mean_pi(self, x: xr.DataArray, **kwargs: Any) -> xr.DataArray:  # noqa: ARG002
+        return xr.dot(self.pi_norm, x, dims=self.dims_n, **self._xarray_dot_kws)  # type: ignore[no-any-return]
 
     @xr_name()
     def mean_pi(
@@ -459,7 +451,7 @@ class xGrandCanonical:
         )
         return self._mean_pi(x, **kwargs)
 
-    # TODO: finish this
+    # TODO(wpk): finish this
     # def _central_moment_bar(
     #     self,
     #     x,
@@ -532,7 +524,7 @@ class xGrandCanonical:
             y = self._array_or_callable_to_xarray(x, **kwargs)
             yy = y - self._mean_pi(y)
 
-        return xr.dot(self.pi_norm, xx, yy, dims=self.dims_n, **self._xarray_dot_kws)  # type: ignore
+        return xr.dot(self.pi_norm, xx, yy, dims=self.dims_n, **self._xarray_dot_kws)  # type: ignore[no-any-return]
 
     def pipe(self, func: Callable[..., R], *args: Any, **kwargs: Any) -> R:
         """Apply function to `self`"""
@@ -585,7 +577,7 @@ class xGrandCanonical:
         # return self.ntot.pipe(lambda x: x / x["volume"])
         return self.ntot / self.volume
 
-    def max(self) -> xr.DataArray:
+    def max(self) -> xr.DataArray:  # noqa: A003
         r""":math:`\max_N \ln \Pi(N)`"""
         return self.lnpi().max(self.dims_n)
 
@@ -606,24 +598,24 @@ class xGrandCanonical:
         x = self.pi_norm.values
         xx = x.reshape(x.shape[0], -1)
         idx_flat = xx.argmax(-1)
-        indexer = np.unravel_index(idx_flat, x.shape[1:])
-        return indexer
+        return np.unravel_index(idx_flat, x.shape[1:])
 
     @cached_prop
     def _argmax_indexer_dict(self) -> dict[str, xr.DataArray]:
         if not isinstance(self._parent, lnPiCollection):
-            raise ValueError("only implemented for lnPiCollection")
+            msg = "only implemented for lnPiCollection"
+            raise ValueError(msg)
 
-        else:
-            return {
-                k: xr.DataArray(v, dims=self._parent._concat_dim)
-                for k, v in zip(self.dims_n, self._argmax_indexer())
-            }
+        return {
+            k: xr.DataArray(v, dims=self._parent._concat_dim)
+            for k, v in zip(self.dims_n, self._argmax_indexer())
+        }
 
     @cached_prop
     def _sample_indexer_dict(self) -> dict[str, xr.DataArray]:
         if not isinstance(self._parent, lnPiCollection):
-            raise ValueError("only implemented for lnPiCollection")
+            msg = "only implemented for lnPiCollection"
+            raise ValueError(msg)
 
         return {
             self._parent._concat_dim: xr.DataArray(
@@ -702,7 +694,7 @@ class xGrandCanonical:
     def _betaOmega(self, lnpi_zero: xArrayLike | None = None) -> xr.DataArray:
         if lnpi_zero is None:
             lnpi_zero = self._lnpi_zero
-        return lnpi_zero - np.log(self.pi_sum)  # type: ignore
+        return lnpi_zero - np.log(self.pi_sum)  # type: ignore[return-value]
 
     def betaOmega(self, lnpi_zero: xArrayLike | None = None) -> xr.DataArray:
         r"""
@@ -781,7 +773,8 @@ class xGrandCanonical:
         """Masks are True where values are stable. Only works for unstacked data."""
 
         if not isinstance(self._parent, lnPiCollection):
-            raise ValueError("only implmented for lnPiCollection")
+            msg = "only implemented for lnPiCollection"
+            raise ValueError(msg)
 
         if not self._xarray_unstack:
             # raise Value'only mask with unstack')
@@ -796,7 +789,7 @@ class xGrandCanonical:
         else:
             out = self.betapV().pipe(lambda x: x.max("phase") == x)
 
-        return out  # type: ignore
+        return out  # type: ignore[no-any-return]
 
     # @cached_meth
     @xr_name(r"$\beta p(\mu,V,T)/\rho$", standard_name="compressibility_factor")
@@ -815,7 +808,7 @@ class xGrandCanonical:
         r"""Potential energy per particle :math:`\overline{PE}/\overline{N}`"""
         return self.PE / self.ntot
 
-    def table(
+    def table(  # noqa: C901, PLR0912
         self,
         keys: str | Sequence[str] | None = None,
         default_keys: str | Sequence[str] | None = ("nvec", "betapV", "PE_n"),
@@ -856,14 +849,14 @@ class xGrandCanonical:
         def _process_keys(x: str | Sequence[str] | None) -> list[str]:
             if x is None:
                 return []
-            elif isinstance(x, str):
+            if isinstance(x, str):
                 return [x]
-            else:
-                return list(x)
+            return list(x)
 
         keys = _process_keys(keys) + _process_keys(default_keys)
         if not keys:
-            raise ValueError("must specify some keys or default_keys to use")
+            msg = "must specify some keys or default_keys to use"
+            raise ValueError(msg)
 
         # only unique keys
         # this preserves order
@@ -871,11 +864,12 @@ class xGrandCanonical:
             try:
                 v = getattr(self, key, None)
                 if v is None:
-                    raise ValueError(f"no attribute {key} in self")
+                    msg = f"no attribute {key} in self"
+                    raise ValueError(msg)
                 if callable(v):
                     v = v()
                 out.append(v)
-            except Exception:
+            except Exception:  # noqa: PERF203, BLE001
                 pass
 
         ds: xr.Dataset = xr.merge(out)
@@ -893,7 +887,7 @@ class xGrandCanonical:
                 ds = (
                     ds.where(mask)
                     .max("phase")
-                    .assign_coords(phase=lambda x: phase[mask.argmax("phase")])
+                    .assign_coords(phase=lambda x: phase[mask.argmax("phase")])  # noqa: ARG005
                 )
 
         if dim_to_suffix is not None:
@@ -909,7 +903,7 @@ class xGrandCanonical:
         r"""Scaled Gibbs free energy :math:`\beta G = \sum_i \beta \mu_i \overline{N}_i`."""
         # return self.betamu.pipe(lambda betamu: (betamu * self.nvec).sum(betamu.dims_comp))
         # return (self.betamu * self.nvec).sum(self.dims_comp)
-        return xr.dot(  # type: ignore
+        return xr.dot(  # type: ignore[no-any-return]
             self.betamu, self.nvec, dims=self.dims_comp, **self._xarray_dot_kws
         )
 
@@ -959,7 +953,7 @@ class xGrandCanonical:
 #     return xCanonical(parent)
 
 
-class xCanonical:
+class xCanonical:  # noqa: N801
     """
     Canonical ensemble properties
 
@@ -1015,7 +1009,7 @@ class xCanonical:
         x = self._parent.xge
 
         if lnpi_zero is None:
-            # TODO: Take another look at this...
+            # TODO(wpk): Take another look at this...
             lnpi_zero = self._parent.data.ravel()[0]
             # lnpi_zero = x.lnpi_zero
 
@@ -1044,7 +1038,8 @@ class xCanonical:
         # if betaPE available, use that:
         PE = self._parent.extra_kws.get("PE", None)
         if PE is None:
-            raise AttributeError('must set "PE" in "extra_kws" of lnPiMasked')
+            msg = 'must set "PE" in "extra_kws" of lnPiMasked'
+            raise AttributeError(msg)
         x = self._xge
         coords = dict(x._wrapper.coords_n, **self._parent.state_kws)
         return xr.DataArray(PE, dims=x.dims_n, coords=coords, attrs=x._standard_attrs)
@@ -1130,9 +1125,7 @@ class xCanonical:
     def table(
         self,
         keys: str | Sequence[str] | None = None,
-        default_keys: str
-        | Sequence[str]
-        | None = (
+        default_keys: str | Sequence[str] | None = (
             "betamu",
             "betapV",
             "PE_n",
@@ -1163,10 +1156,9 @@ class xCanonical:
         def _process_keys(x: str | Sequence[str] | None) -> list[str]:
             if x is None:
                 return []
-            elif isinstance(x, str):
+            if isinstance(x, str):
                 return [x]
-            else:
-                return list(x)
+            return list(x)
 
         keys_total = _process_keys(keys) + _process_keys(default_keys)
 
@@ -1177,7 +1169,7 @@ class xCanonical:
                 if callable(v):
                     v = v()
                 out.append(v)
-            except Exception:
+            except Exception:  # noqa: PERF203, BLE001
                 pass
 
         ds = xr.merge(out)
