@@ -39,29 +39,26 @@ def _get_shift(
     shape: tuple[int, ...], dlnz: tuple[float, ...], dtype: DTypeLike
 ) -> MyNDArray:
     shift = np.zeros([], dtype=dtype)
-    for i, (nr, m) in enumerate(zip(_get_n_ranges(shape=shape, dtype=dtype), dlnz)):  # type: ignore
-        shift = np.add.outer(shift, nr * m)  # pyright: ignore
+    for _i, (nr, m) in enumerate(zip(_get_n_ranges(shape=shape, dtype=dtype), dlnz)):  # type: ignore[arg-type]
+        shift = np.add.outer(shift, nr * m)
     return shift
 
 
 @lru_cache(maxsize=20)
 def _get_data(base: lnPiArray, dlnz: tuple[float, ...]) -> MyNDArray:
     if all(x == 0 for x in dlnz):
-        return base.data  # pyright: ignore
-    else:
-        return (
-            _get_shift(base.shape, dlnz, base.data.dtype) + base.data
-        )  # pyright: ignore
+        return base.data
+    return _get_shift(base.shape, dlnz, base.data.dtype) + base.data
 
 
 @lru_cache(maxsize=20)
 def _get_maskedarray(
     base: lnPiArray, self: lnPiMasked, dlnz: tuple[float, ...]
 ) -> np.ma.core.MaskedArray[Any, np.dtype[Any]]:
-    return np.ma.MaskedArray(  # type: ignore
+    return np.ma.MaskedArray(  # type: ignore[no-untyped-call]
         _get_data(base, dlnz),
         mask=self._mask,
-        fill_value=base.fill_value,  # pyright: ignore
+        fill_value=base.fill_value,
     )
 
 
@@ -72,10 +69,10 @@ def _get_filled(
     dlnz: tuple[float, ...],
     fill_value: float | None = None,
 ) -> np.ma.core.MaskedArray[Any, np.dtype[Any]]:
-    return _get_maskedarray(base, self, dlnz).filled(fill_value)  # type: ignore
+    return _get_maskedarray(base, self, dlnz).filled(fill_value)  # type: ignore[no-any-return, no-untyped-call]
 
 
-class lnPiArray:
+class lnPiArray:  # noqa: N801
     """
     Wrapper on lnPi lnPiArray
 
@@ -190,7 +187,7 @@ class lnPiArray:
             object with padded data
         """
 
-        import bottleneck  # pyright: ignore
+        import bottleneck
 
         from . import utils
 
@@ -210,8 +207,7 @@ class lnPiArray:
         if len(datas) > 0:
             data = bottleneck.nanmean(datas, axis=0)
 
-        new = self.new_like(data=data)
-        return new
+        return self.new_like(data=data)
 
     def zeromax(self, mask: MyNDArray | bool = False) -> Self:
         """
@@ -224,12 +220,12 @@ class lnPiArray:
             data is excluded from calculating maximum.
         """
 
-        data = self.data - np.ma.MaskedArray(self.data, mask).max()  # type: ignore
+        data = self.data - np.ma.MaskedArray(self.data, mask).max()  # type: ignore[no-untyped-call]
         return self.new_like(data=data)
 
 
-@docfiller.decorate
-class lnPiMasked(AccessorMixin):
+@docfiller.decorate  # noqa: PLR0904
+class lnPiMasked(AccessorMixin):  # noqa: N801
     """
     Masked array like wrapper for lnPi data.
 
@@ -339,7 +335,7 @@ class lnPiMasked(AccessorMixin):
 
     @property
     def _data(self) -> MyNDArray:
-        return self._base.data  # pyright: ignore
+        return self._base.data
 
     @property
     def dtype(self) -> np.dtype[Any]:
@@ -403,12 +399,12 @@ class lnPiMasked(AccessorMixin):
     @property
     def volume(self) -> float | None:
         """Accessor to self.state_kws['volume']."""
-        return self.state_kws.get("volume", None)
+        return self.state_kws.get("volume", None)  # type: ignore[no-any-return]
 
     @property
     def beta(self) -> float | None:
         """Accessor to self.state_kws['beta']."""
-        return self.state_kws.get("beta", None)
+        return self.state_kws.get("beta", None)  # type: ignore[no-any-return]
 
     def __repr__(self) -> str:
         return f"<lnPi(lnz={self._lnz})>"
@@ -462,7 +458,7 @@ class lnPiMasked(AccessorMixin):
         numpy.ma.MaskedArray.argmax
         numpy.unravel_index
         """
-        return np.unravel_index(self.ma.argmax(*args, **kwargs), self.shape)  # type: ignore
+        return np.unravel_index(self.ma.argmax(*args, **kwargs), self.shape)  # type: ignore[return-value,no-untyped-call]
 
     # @cached.meth
     def local_max(
@@ -482,14 +478,14 @@ class lnPiMasked(AccessorMixin):
         --------
         numpy.ma.MaskedArray.max
         """
-        return self.ma[self.local_argmax(*args, **kwargs)]  # type: ignore
+        return self.ma[self.local_argmax(*args, **kwargs)]  # type: ignore[no-any-return]
 
     # @cached.meth
     def local_maxmask(
         self, *args: Any, **kwargs: Any
     ) -> np.ma.core.MaskedArray[Any, np.dtype[Any]]:
         """Calculate mask where ``self.ma == self.local_max()``"""
-        return self.ma == self.local_max(*args, **kwargs)  # type: ignore
+        return self.ma == self.local_max(*args, **kwargs)  # type: ignore[no-any-return]
 
     @cached.prop
     def edge_distance_matrix(self) -> NDArray[np.float_]:
@@ -518,7 +514,7 @@ class lnPiMasked(AccessorMixin):
         edge_distance_matrix
         lnpy.utils.distance_matrix
         """
-        return ref.edge_distance_matrix[self.local_argmax(*args, **kwargs)]  # type: ignore
+        return ref.edge_distance_matrix[self.local_argmax(*args, **kwargs)]  # type: ignore[no-any-return]
 
     @docfiller.decorate
     def new_like(
@@ -596,11 +592,11 @@ class lnPiMasked(AccessorMixin):
         """Create new object at specified value of `lnz`"""
         return self.new_like(lnz=lnz)
 
-    def or_mask(self, mask: MyNDArray, **kwargs: Any) -> Self:
+    def or_mask(self, mask: MyNDArray) -> Self:
         """New object with logical or of self.mask and mask"""
         return self.new_like(mask=(mask | self.mask))
 
-    def and_mask(self, mask: MyNDArray, **kwargs: Any) -> Self:
+    def and_mask(self, mask: MyNDArray) -> Self:
         """New object with logical and of self.mask and mask"""
         return self.new_like(mask=(mask & self.mask))
 
@@ -644,7 +640,7 @@ class lnPiMasked(AccessorMixin):
             csv_kws = {}
 
         da = (
-            pd.read_csv(path, sep=sep, names=names, **csv_kws)  # type: ignore
+            pd.read_csv(path, sep=sep, names=names, **csv_kws)  # type: ignore[call-overload]
             .set_index(names[:-1])["lnpi"]
             .to_xarray()
         )
@@ -694,11 +690,8 @@ class lnPiMasked(AccessorMixin):
         # where are state variables
         if state_as_attrs is None:
             state_as_attrs = bool(da.attrs.get("state_as_attrs", False))
-        if state_as_attrs:
-            # state variables from attrs
-            c = da.attrs
-        else:
-            c = da.coords  # type: ignore [assignment]
+
+        c = da.attrs if state_as_attrs else da.coords
 
         lnz = []
         state_kws = {}
