@@ -28,6 +28,7 @@ if TYPE_CHECKING:
     from typing import (
         Any,
         Callable,
+        Final,
         Hashable,
         Iterable,
         Iterator,
@@ -44,7 +45,7 @@ if TYPE_CHECKING:
     from ._typing import IndexingInt, MyNDArray, Scalar
 
 
-class SeriesWrapper(AccessorMixin, Generic[T_Element]):
+class SeriesWrapper(AccessorMixin, Generic[T_Element]):  # noqa: PLR0904
     """wrap object in series"""
 
     def __init__(
@@ -62,7 +63,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         self._base_class = base_class
         self._verify = self._base_class is not None
 
-        series = pd.Series(data=data, index=index, dtype=dtype, name=name)  # type: ignore
+        series = pd.Series(data=data, index=index, dtype=dtype, name=name)  # type: ignore[misc,arg-type]
         self._verify_series(series)
         self._series = series
         self._cache: dict[str, Any] = {}
@@ -76,7 +77,8 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
 
             for d in series:
                 if not issubclass(type(d), base_class):
-                    raise ValueError(f"all elements must be of type {base_class}")
+                    msg = f"all elements must be of type {base_class}"
+                    raise ValueError(msg)
 
     @property
     def series(self) -> pd.Series[Any]:
@@ -100,7 +102,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
     @property
     def values(self) -> MyNDArray:
         """Series values"""
-        return self._series.values  # type: ignore
+        return self._series.values  # type: ignore[return-value]
 
     @property
     def items(self) -> MyNDArray:
@@ -133,7 +135,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         return type(self)(
             data=data,
             index=index,
-            dtype=self.s.dtype,  # type: ignore
+            dtype=self.s.dtype,  # type: ignore[arg-type]
             name=self.s.name,
             base_class=self._base_class,
             **kwargs,
@@ -146,11 +148,11 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         val = getattr(self._series, mtd)(*args, **kwargs)
         if wrap and type(val) == pd.Series:
             val = self.new_like(val)
-        return val  # type: ignore
+        return val  # type: ignore[no-any-return]
 
     def __getitem__(self, key: Any) -> Self | T_Element:
         """Interface to :meth:`pandas.Series.__getitem__`"""
-        return self._wrapped_pandas_method("__getitem__", wrap=True, key=key)  # type: ignore
+        return self._wrapped_pandas_method("__getitem__", wrap=True, key=key)  # type: ignore[return-value]
 
     def xs(
         self,
@@ -172,7 +174,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         self._series[idx] = values
 
     def __repr__(self) -> str:
-        return f"<class {self.__class__.__name__}>\n{repr(self.s)}"
+        return f"<class {self.__class__.__name__}>\n{self.s!r}"
 
     def __str__(self) -> str:
         return str(self.s)
@@ -211,7 +213,8 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         elif isinstance(to_append, pd.Series):
             series = to_append
         else:
-            raise ValueError(f"Unknown to append type={type(to_append)}")
+            msg = f"Unknown to append type={type(to_append)}"
+            raise ValueError(msg)
 
         if concat_kws is None:
             concat_kws = {}
@@ -226,8 +229,8 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         if inplace:
             self.series = s
             return self
-        else:
-            return self.new_like(s)
+
+        return self.new_like(s)
 
     def droplevel(self, level: int | Hashable | Sequence[int | Hashable]) -> Self:
         """
@@ -237,7 +240,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         --------
         pandas.Series.droplevel
         """
-        return self.new_like(self._series.droplevel(level=level, axis=0))  # type: ignore
+        return self.new_like(self._series.droplevel(level=level, axis=0))  # type: ignore[arg-type,unused-ignore]
 
     def apply(
         self,
@@ -249,7 +252,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
     ) -> Self | pd.Series[Any]:
         """Interface to :meth:`pandas.Series.apply`"""
 
-        return self._wrapped_pandas_method(  # type: ignore
+        return self._wrapped_pandas_method(  # type: ignore[return-value]
             "apply",
             wrap=wrap,
             func=func,
@@ -260,7 +263,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
 
     def sort_index(self, *args: Any, **kwargs: Any) -> Self:
         """Interface to :meth:`pandas.Series.sort_index`"""
-        return self._wrapped_pandas_method("sort_index", *args, wrap=True, **kwargs)  # type: ignore
+        return self._wrapped_pandas_method("sort_index", *args, wrap=True, **kwargs)  # type: ignore[misc,return-value]
 
     @overload
     def groupby(
@@ -309,7 +312,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
 
     def groupby(
         self,
-        by: Hashable | Sequence[Hashable] = None,
+        by: Hashable | Sequence[Hashable] | None = None,
         *,
         level: int | Hashable | Sequence[int | Hashable] | None = None,
         as_index: bool = True,
@@ -332,8 +335,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         --------
         pandas.Series.groupby
         """
-        # TODO: fix types
-        group = self.s.groupby(  # type: ignore
+        group = self.s.groupby(  # type: ignore[call-overload]
             by=by,
             axis=0,
             level=level,
@@ -347,8 +349,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
 
         if wrap:
             return _Groupby(self, group)
-        else:
-            return group  # type: ignore
+        return group  # type: ignore[no-any-return]
 
     @overload
     def groupby_allbut(
@@ -400,8 +401,7 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
         # To suppress annoying errors.
         if len(by) == 1:
             return self.groupby(by=by[0], wrap=wrap, **kwargs)
-        else:
-            return self.groupby(by=by, wrap=wrap, **kwargs)
+        return self.groupby(by=by, wrap=wrap, **kwargs)
 
     @classmethod
     def _concat_to_series(
@@ -424,18 +424,16 @@ class SeriesWrapper(AccessorMixin, Generic[T_Element]):
             for k in objs:
                 v = objs[k]
                 if remap is None:
-                    if isinstance(v, cls):
-                        remap = True
-                    else:
-                        remap = False
+                    remap = bool(isinstance(v, cls))
                 if remap:
                     out[k] = v._series
                 else:
                     out[k] = v
             objs = out
         else:
-            raise ValueError(f"bad input type {type(objs[0])}")
-        return pd.concat(objs, **concat_kws)  # type: ignore
+            msg = f"bad input type {type(objs[0])}"
+            raise ValueError(msg)
+        return pd.concat(objs, **concat_kws)  # type: ignore[return-value,arg-type]
 
     def concat_like(
         self,
@@ -522,10 +520,10 @@ class _Groupby(Generic[T_SeriesWrapper, T_Element]):
             out = getattr(self._group, attr)
             if callable(out):
                 return _CallableResult(self._parent, out)
-            else:
-                return self._parent.new_like(out)
-        else:
-            raise AttributeError(f"no attribute {attr} in groupby")
+            return self._parent.new_like(out)
+
+        msg = f"no attribute {attr} in groupby"
+        raise AttributeError(msg)
 
 
 # @SeriesWrapper.decorate_accessor("loc")
@@ -559,7 +557,7 @@ class _LocIndexer(Generic[T_SeriesWrapper, T_Element]):
         out = self._loc[idx]
         if isinstance(out, pd.Series):
             out = self._parent.new_like(out)
-        return out  # type: ignore
+        return out  # type: ignore[no-any-return]
 
     def __setitem__(
         self, idx: Any, values: T_Element | pd.Series[Any] | Sequence[T_Element]
@@ -568,7 +566,7 @@ class _LocIndexer(Generic[T_SeriesWrapper, T_Element]):
 
 
 # @SeriesWrapper.decorate_accessor("iloc")
-class _iLocIndexer(Generic[T_SeriesWrapper, T_Element]):
+class _iLocIndexer(Generic[T_SeriesWrapper, T_Element]):  # noqa: N801
     """
     Indexer by position.
 
@@ -597,7 +595,7 @@ class _iLocIndexer(Generic[T_SeriesWrapper, T_Element]):
         out = self._iloc[idx]
         if isinstance(out, pd.Series):
             out = self._parent.new_like(out)
-        return out  # type: ignore
+        return out  # type: ignore[no-any-return]
 
     def __setitem__(
         self, idx: Any, values: T_Element | pd.Series[Any] | Sequence[T_Element]
@@ -619,11 +617,11 @@ class _Query(Generic[T_SeriesWrapper, T_Element]):
 
     def __call__(self, expr: str, **kwargs: Any) -> T_SeriesWrapper:
         idx = self._frame.query(expr, **kwargs).index
-        return self._parent.iloc[idx]  # type: ignore
+        return self._parent.iloc[idx]  # type: ignore[no-any-return]
 
 
 # @SeriesWrapper.decorate_accessor("zloc")
-class _LocIndexer_unstack_zloc(Generic[T_SeriesWrapper, T_Element]):
+class _LocIndexer_unstack_zloc(Generic[T_SeriesWrapper, T_Element]):  # noqa: N801
     """positional indexer for everything but phase"""
 
     def __init__(
@@ -637,20 +635,18 @@ class _LocIndexer_unstack_zloc(Generic[T_SeriesWrapper, T_Element]):
 
     def __getitem__(self, idx: Any) -> T_SeriesWrapper:
         out = self._loc[idx]
-        if isinstance(out, pd.DataFrame):
-            out = out.stack(self._level)
-        else:
-            out = out.dropna()
+        out = out.stack(self._level) if isinstance(out, pd.DataFrame) else out.dropna()
 
         if isinstance(out, pd.Series):
             out = self._parent.new_like(out)
         else:
-            raise ValueError("unknown indexer for zloc")
-        return out  # type: ignore
+            msg = "unknown indexer for zloc"
+            raise ValueError(msg)
+        return out  # type: ignore[no-any-return]
 
 
 # @SeriesWrapper.decorate_accessor("mloc")
-class _LocIndexer_unstack_mloc(Generic[T_SeriesWrapper, T_Element]):
+class _LocIndexer_unstack_mloc(Generic[T_SeriesWrapper, T_Element]):  # noqa: N801
     """indexer with pandas index"""
 
     def __init__(
@@ -672,12 +668,11 @@ class _LocIndexer_unstack_mloc(Generic[T_SeriesWrapper, T_Element]):
             drop: list[Hashable] = list(self._index_names - set(idx.names))
             index = index.droplevel(drop)
             # reorder idx
-            idx = idx.reorder_levels(index.names)  # type: ignore
+            idx = idx.reorder_levels(index.names)  # type: ignore[no-untyped-call]
         else:
             drop = list(set(index.names) - {idx.name})
             index = index.droplevel(drop)
-        indexer = index.get_indexer_for(idx)  # type: ignore
-        return indexer
+        return index.get_indexer_for(idx)  # type: ignore[no-untyped-call]
 
     def __getitem__(self, idx: pd.MultiIndex | pd.Index[Any]) -> T_SeriesWrapper:
         indexer = self._get_loc_idx(idx)
@@ -686,11 +681,12 @@ class _LocIndexer_unstack_mloc(Generic[T_SeriesWrapper, T_Element]):
         if isinstance(out, pd.Series):
             out = self._parent.new_like(out)
         else:
-            raise ValueError("unknown indexer for mloc")
-        return out  # type: ignore
+            msg = "unknown indexer for mloc"
+            raise ValueError(msg)
+        return out  # type: ignore[no-any-return]
 
 
-class lnPiCollection(SeriesWrapper[lnPiMasked]):
+class lnPiCollection(SeriesWrapper[lnPiMasked]):  # noqa: N801
     # class lnPiCollection:
     r"""
     Wrapper around :class:`pandas.Series` for collection of :class:`~lnpy.lnpidata.lnPiMasked` objects.
@@ -724,7 +720,7 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
     _use_joblib = True
     _xarray_output = True
     _xarray_unstack = True
-    _xarray_dot_kws = {"optimize": "optimal"}
+    _xarray_dot_kws: Final = {"optimize": "optimal"}
     _use_cache = True
 
     def __init__(
@@ -787,10 +783,10 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
     # repr
     @cached.prop
     def _lnz_series(self) -> pd.Series[Any]:
-        return self._series.apply(lambda x: x.lnz)  # type: ignore
+        return self._series.apply(lambda x: x.lnz)  # type: ignore[no-any-return]
 
     def __repr__(self) -> str:
-        return f"<class {self.__class__.__name__}>\n{repr(self._lnz_series)}"
+        return f"<class {self.__class__.__name__}>\n{self._lnz_series!r}"
 
     def __str__(self) -> str:
         return str(self._lnz_series)
@@ -820,13 +816,12 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
             .to_frame()
             .assign(lnz_sample=lambda x: np.arange(len(x)))["lnz_sample"]
         )
-        index_frame = (
+        return (
             self.index.to_frame()
             .reset_index("phase", drop=True)[["phase"]]
             .assign(lnz_index=lambda x: sample_frame[x.index])
             .reset_index()
         )
-        return index_frame
 
     @overload
     def _get_lnz(
@@ -847,14 +842,11 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
         Helper function to.
         returns self.iloc[idx].lnz[component]
         """
-        if zloc is not None:
-            s = self.zloc[zloc]._series
-        else:
-            s = self._series
+        s = self.zloc[zloc]._series if zloc is not None else self._series
         lnz = s.iloc[iloc].lnz
         if component is not None:
             lnz = lnz[component]
-        return lnz  # type: ignore
+        return lnz  # type: ignore[no-any-return]
 
     def _get_level(self, level: str = "phase") -> pd.Index[Any]:
         """Return level values from index"""
@@ -882,7 +874,7 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
         # but makes clear where the time is being spent
         first: lnPiMasked = self.iloc[0]
         n = len(self)
-        out = np.empty((n,) + first.shape, dtype=first.dtype)
+        out = np.empty((n, *first.shape), dtype=first.dtype)
         seq = get_tqdm((x.filled(fill_value) for x in self), total=n)
         for i, x in enumerate(seq):
             out[i, ...] = x
@@ -906,7 +898,7 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
         first = self.iloc[0]
         n = len(self)
 
-        pi_norm = np.empty((n,) + first.shape, dtype=first.dtype)
+        pi_norm = np.empty((n, *first.shape), dtype=first.dtype)
         pi_sum = np.empty(n, dtype=first.dtype)
         lnpi_zero = np.empty(n, dtype=first.dtype)
 
@@ -925,9 +917,8 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
     ) -> Sequence[Any] | xr.DataArray:
         """Utility to wrap output in :class:xarray.DataArray"""
         if self._xarray_output and isinstance(items[0], xr.DataArray):
-            return xr.concat(items, self.index, coords=self._concat_coords)  # type: ignore
-        else:
-            return items
+            return xr.concat(items, self.index, coords=self._concat_coords)  # type: ignore[call-overload,no-any-return]
+        return items
 
     ##################################################
     # Constructors
@@ -967,14 +958,13 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
     def from_builder(
         cls,
         lnzs: Sequence[float] | MyNDArray,
-        # TODO: make better type for build_phases.
+        # TODO(wpk): make better type for build_phases.
         build_phases: Callable[..., tuple[list[lnPiMasked], MyNDArray]],
         ref: lnPiMasked | None = None,
         build_kws: Mapping[str, Any] | None = None,
         nmax: int | None = None,
         # concat_kws: Mapping[str, Any] |None=None,
         base_class: str | type = "first",
-        *args: Any,
         **kwargs: Any,
     ) -> Self:
         """
@@ -1008,16 +998,15 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
 
         build_kws = dict(build_kws, phases_factory=False)
         seq = get_tqdm(lnzs, desc="build")
-        L = parallel_map(build_phases, seq, ref=ref, nmax=nmax, **build_kws)
-        # return cls.concat(L, verify=verify, concat_kws=concat_kws, base_class=base_class,
-        #                   *args, **kwargs)
 
         items: list[lnPiMasked] = []
         index: list[int] = []
-        for data, idx in L:
+        for data, idx in parallel_map(
+            build_phases, seq, ref=ref, nmax=nmax, **build_kws
+        ):
             items += data
             index += list(idx)
-        return cls.from_list(items, index, base_class=base_class, *args, **kwargs)
+        return cls.from_list(items, index, base_class=base_class, **kwargs)
 
     ################################################################################
     # dataarray io
@@ -1048,7 +1037,7 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
                 masks_to_labels(masks, features=features, convention=False, dtype=dtype)
             )
 
-        index = indexes[0].append(indexes[1:])  # type: ignore
+        index = indexes[0].append(indexes[1:])  # type: ignore[no-untyped-call]
 
         data = np.stack(labels)
 
@@ -1205,23 +1194,29 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
         return wFreeEnergyPhases(self)
 
     @property
-    def wlnPi(self) -> lnpienergy.wFreeEnergyCollection:
+    def wlnPi(self) -> lnpienergy.wFreeEnergyCollection:  # noqa: N802
         """
         Deprecated accessor to :class:`~lnpy.lnpienergy.wFreeEnergyCollection` from :attr:`wlnPi`.
 
         Alias to :attr:`wfe`
         """
-        warn("Using `wlnPi` accessor is deprecated.  Please use `wfe` accessor instead")
+        warn(
+            "Using `wlnPi` accessor is deprecated.  Please use `wfe` accessor instead",
+            stacklevel=1,
+        )
         return self.wfe
 
     @property
-    def wlnPi_single(self) -> lnpienergy.wFreeEnergyPhases:
+    def wlnPi_single(self) -> lnpienergy.wFreeEnergyPhases:  # noqa: N802
         """
         Deprecated accessor to :class:`~lnpy.lnpienergy.wFreeEnergyPhases` from :attr:`wlnPi_single`.
 
         Alias to :attr:`wfe_phases`
         """
-        warn("Using `wlnPi_single is deprecated.  Please use `self.wfe_phases` instead")
+        warn(
+            "Using `wlnPi_single is deprecated.  Please use `self.wfe_phases` instead",
+            stacklevel=1,
+        )
         return self.wfe_phases
 
     @cached.prop
@@ -1246,20 +1241,21 @@ class lnPiCollection(SeriesWrapper[lnPiMasked]):
         copy_stability: bool = True,
     ) -> Self:
         if (not append) and (not copy_stability):
-            raise ValueError("one of append or copy_stability must be True")
+            msg = "one of append or copy_stability must be True"
+            raise ValueError(msg)
 
         if other is None:
             other = self
         spin = other.spinodal
         bino = other.binodal
         if append:
-            new = self.append(spin.appender).append(bino.appender)  # type: ignore
+            new = self.append(spin.appender).append(bino.appender)  # type: ignore[arg-type]
             if sort:
                 new = new.sort_index()
         else:
             new = self.copy()
         if copy_stability:
-            # TODO: fix this hack
+            # TODO(wpk): fix this hack
             new._cache["spinodal"] = spin
             new._cache["binodal"] = bino
             # new.spinodal = spin
