@@ -19,9 +19,9 @@ if TYPE_CHECKING:
     from .segment import BuildPhasesBase
 
 
-def _initial_bracket_molfrac(
+def _initial_bracket_molfrac(  # noqa: C901,PLR0912,PLR0915
     target: float,
-    C: lnPiCollection,
+    collection: lnPiCollection,
     build_phases: BuildPhasesBase,
     phase_id: int | str | None = 0,
     component: int | None = None,
@@ -47,8 +47,7 @@ def _initial_bracket_molfrac(
             v = x.xge.molfrac.sel(component=component)
             if len(x) == 1:
                 return v.isel(phase=0)
-            else:
-                return v.where(x.xge.mask_stable).max("phase")
+            return v.where(x.xge.mask_stable).max("phase")
 
     else:
         skip_phase_id = False
@@ -56,22 +55,19 @@ def _initial_bracket_molfrac(
         def getter(x: lnPiCollection) -> xr.DataArray:
             return x.xge.molfrac.sel(component=component, phase=phase_id)
 
-    s = getter(C).to_series().dropna()
+    s = getter(collection).to_series().dropna()
 
     # get left bound
     left = None
     ntry_left = 0
     ss = s[s < target]
     if len(ss) > 0:
-        left = C.mloc[ss.index[[-1]]]
+        left = collection.mloc[ss.index[[-1]]]
     else:
-        if len(s) > 0:
-            index = s.index[[0]]
-        else:
-            index = C.index[[0]].droplevel("phase")
-        new_lnz = C.mloc[index]._get_lnz(lnz_idx)
+        index = s.index[[0]] if len(s) > 0 else collection.index[[0]].droplevel("phase")
+        new_lnz = collection.mloc[index]._get_lnz(lnz_idx)
         dlnz_ = dlnz
-        for i in range(ntry):
+        for i in range(ntry):  # noqa: B007
             new_lnz -= dlnz_
             dlnz_ *= dfac
             p = build_phases(new_lnz, ref=ref, **build_kws)
@@ -83,23 +79,23 @@ def _initial_bracket_molfrac(
         ntry_left = i
 
     if left is None:
-        raise RuntimeError("could not find left bounds")
+        msg = "could not find left bounds"
+        raise RuntimeError(msg)
 
     # right bracket
     right = None
     ntry_right = 0
     ss = s[s > target]
     if len(ss) > 0:
-        right = C.mloc[ss.index[[0]]]
+        right = collection.mloc[ss.index[[0]]]
     else:
-        if len(s) > 0:
-            index = s.index[[-1]]
-        else:
-            index = C.index[[-1]].droplevel("phase")
-        new_lnz = C.mloc[index]._get_lnz(lnz_idx)
+        index = (
+            s.index[[-1]] if len(s) > 0 else collection.index[[-1]].droplevel("phase")
+        )
+        new_lnz = collection.mloc[index]._get_lnz(lnz_idx)
         dlnz_ = dlnz
 
-        for i in range(ntry):
+        for i in range(ntry):  # noqa: B007
             new_lnz += dlnz_
             p = build_phases(new_lnz, ref=ref, **build_kws)
             if (not skip_phase_id) and (phase_id not in p._get_level("phase")):
@@ -115,13 +111,14 @@ def _initial_bracket_molfrac(
         ntry_right = i
 
     if right is None:
-        raise RuntimeError("could not find right bounds")
+        msg = "could not find right bounds"
+        raise RuntimeError(msg)
 
-    info = dict(ntry_left=ntry_left, ntry_right=ntry_right)
+    info = {"ntry_left": ntry_left, "ntry_right": ntry_right}
     return left, right, info
 
 
-def _solve_lnz_molfrac(
+def _solve_lnz_molfrac(  # noqa: C901
     target: float,
     left: lnPiCollection | float,
     right: lnPiCollection | float,
@@ -174,7 +171,7 @@ def _solve_lnz_molfrac(
     if not isinstance(right, float):
         right = right._get_lnz(build_phases.index)
 
-    a, b = sorted([x for x in (left, right)])
+    a, b = sorted((left, right))
 
     if isinstance(phase_id, str) and phase_id.lower() == "none":
         # select stable phase
@@ -184,8 +181,7 @@ def _solve_lnz_molfrac(
             v = x.xge.molfrac.sel(component=component)
             if len(x) == 1:
                 return v.isel(phase=0)
-            else:
-                return v.where(x.xge.mask_stable).max("phase")
+            return v.where(x.xge.mask_stable).max("phase")
 
     else:
         skip_phase_id = False
@@ -195,10 +191,10 @@ def _solve_lnz_molfrac(
 
     def f(x: float) -> float | MyNDArray:
         p = build_phases(x, ref=ref, **build_kws)
-        f.lnpi = p  # type: ignore
+        f.lnpi = p  # type: ignore[attr-defined]
 
         # by not using the ListAccessor,
-        # can parralelize
+        # can parallelize
         mf: MyNDArray | float
         if skip_phase_id or phase_id in p._get_level("phase"):
             mf = getter(p).values
@@ -218,14 +214,15 @@ def _solve_lnz_molfrac(
 
     residual = f(xx)
     if np.abs(residual) > tol:
-        raise RuntimeError("something went wrong with solve")
+        msg = "something went wrong with solve"
+        raise RuntimeError(msg)
 
-    return f.lnpi, rootresults_to_rootresultdict(r, residual=residual)  # type: ignore
+    return f.lnpi, rootresults_to_rootresultdict(r, residual=residual)  # type: ignore[attr-defined]
 
 
 def find_lnz_molfrac(
     target: float,
-    C: lnPiCollection,
+    collection: lnPiCollection,
     build_phases: BuildPhasesBase,
     phase_id: int | str | None = 0,
     component: int | None = None,
@@ -237,9 +234,9 @@ def find_lnz_molfrac(
     tol: float = 1e-4,
     **kwargs: Any,
 ) -> tuple[lnPiCollection, RootResultDict]:
-    left, right, info = _initial_bracket_molfrac(
+    left, right, _info = _initial_bracket_molfrac(
         target=target,
-        C=C,
+        collection=collection,
         build_phases=build_phases,
         phase_id=phase_id,
         component=component,
