@@ -147,7 +147,7 @@ def peak_local_max_adaptive(
 
 
 @docfiller_local
-def peak_local_max_adaptive(
+def peak_local_max_adaptive(  # noqa: C901
     data: MyNDArray,
     *,
     mask: MyNDArray | None = None,
@@ -201,7 +201,7 @@ def peak_local_max_adaptive(
     from skimage.feature import peak_local_max
     from skimage.morphology import label as morphology_label
 
-    assert style in ["indices", "mask", "marker"]
+    assert style in {"indices", "mask", "marker"}
 
     if min_distance is None:
         min_distance = [5, 10, 15, 20, 25]
@@ -213,7 +213,7 @@ def peak_local_max_adaptive(
         min_distance = [min_distance]
 
     data = data - bottleneck.nanmin(data)
-    kwargs = dict(dict(exclude_border=False), **kwargs)
+    kwargs = dict({"exclude_border": False}, **kwargs)
 
     n = idx = None
     for md in min_distance:
@@ -232,33 +232,31 @@ def peak_local_max_adaptive(
             break
 
     if n is None or idx is None:
-        raise ValueError("failed to find peaks")
+        msg = "failed to find peaks"
+        raise ValueError(msg)
 
     if n > num_peaks_max:
         if errors == "ignore":
             pass
-        elif errors in ("raise", "ignore"):
+        elif errors in {"raise", "ignore"}:
             message = f"{n} maxima found greater than {num_peaks_max}"
             if errors == "raise":
                 raise RuntimeError(message)
-            else:
-                warnings.warn(message)
+            warnings.warn(message, stacklevel=1)
 
     idx = tuple(idx.T)
     if style == "indices":
         return cast("tuple[MyNDArray, ...]", idx)
 
-    else:
-        out = np.zeros_like(data, dtype=bool)
-        out[idx] = True
+    out = np.zeros_like(data, dtype=bool)
+    out[idx] = True
 
-        if style == "marker":
-            out = morphology_label(out, connectivity=connectivity)
-
-        return cast("MyNDArray", out)
+    if style == "marker":
+        out = morphology_label(out, connectivity=connectivity)
+    return cast("MyNDArray", out)
 
 
-# TODO: create these for below
+# TODO(wpk): create these for below
 # class Peak_Kws(TypedDict, total=False):
 #     min_distance: Sequence[int]
 #     style: PeakStyle
@@ -396,9 +394,7 @@ class Segmenter:
             kwargs["num_peaks_max"] = num_peaks_max
         kwargs["style"] = style
         kwargs = dict(self.peak_kws, **kwargs)
-        out = peak_local_max_adaptive(data, **kwargs)
-
-        return out  # type: ignore
+        return peak_local_max_adaptive(data, **kwargs)  # type: ignore[no-any-return]
 
     @docfiller_local
     def watershed(
@@ -434,8 +430,8 @@ class Segmenter:
         if connectivity is None:
             connectivity = data.ndim
 
-        kwargs = dict(self.watershed_kws, connectivity=connectivity, *kwargs)
-        return watershed(data, markers=markers, mask=mask, **kwargs)  # type: ignore
+        kwargs = dict(self.watershed_kws, connectivity=connectivity, **kwargs)
+        return watershed(data, markers=markers, mask=mask, **kwargs)  # type: ignore[no-any-return]
 
     @docfiller_local
     def segment_lnpi(
@@ -499,10 +495,9 @@ class Segmenter:
 
         if watershed_kws is None:
             watershed_kws = {}
-        labels = self.watershed(
+        return self.watershed(
             -lnpi.data, markers=markers, mask=~lnpi.mask, connectivity=connectivity
         )
-        return labels
 
 
 class PhaseCreator:
@@ -528,7 +523,7 @@ class PhaseCreator:
     phases_factory : callable, optional
         Factory function for returning Collection from a list of :class:`~lnpy.lnpidata.lnPiMasked` object.
         Defaults to :meth:`.lnPiCollection.from_list`.
-    lnPiFreeEnergy_kws : mapping, optional
+    free_energy_kws : mapping, optional
         Optional arguments to ...
     merge_kws : mapping, optional
         Optional arguments to :func:`.merge_regions`
@@ -544,7 +539,7 @@ class PhaseCreator:
         segment_kws: Mapping[str, Any] | None = None,
         tag_phases: TagPhasesSignature | None = None,
         phases_factory: PhasesFactorySignature | None = None,
-        lnPiFreeEnergy_kws: Mapping[str, Any] | None = None,
+        free_energy_kws: Mapping[str, Any] | None = None,
         merge_kws: Mapping[str, Any] | None = None,
     ) -> None:
         self.nmax = nmax
@@ -557,16 +552,16 @@ class PhaseCreator:
         self.segment_kws = {} if segment_kws is None else dict(segment_kws)
         self.segment_kws["num_peaks_max"] = nmax_peak or nmax * 2
 
-        self.lnPiFreeEnergy_kws = lnPiFreeEnergy_kws or {}
+        self.free_energy_kws = free_energy_kws or {}
         self.merge_kws = (
             {}
             if merge_kws is None
             else dict(merge_kws, convention=False, nfeature_max=self.nmax)
         )
 
-    # TODO: make this work with integer or string phase_ids
+    # TODO(wpk): make this work with integer or string phase_ids
+    @staticmethod
     def _merge_phase_ids(
-        self,
         ref: lnPiMasked,
         phase_ids: Sequence[int] | MyNDArray,
         lnpis: list[lnPiMasked],
@@ -612,7 +607,7 @@ class PhaseCreator:
         phases_factory: PhasesFactorySignature | Literal[True] = ...,
         phase_kws: Mapping[str, Any] | None = ...,
         segment_kws: Mapping[str, Any] | None = ...,
-        lnPiFreeEnergy_kws: Mapping[str, Any] | None = ...,
+        free_energy_kws: Mapping[str, Any] | None = ...,
         merge_kws: Mapping[str, Any] | None = ...,
         tag_phases: TagPhasesSignature | None = ...,
     ) -> lnPiCollection:
@@ -634,7 +629,7 @@ class PhaseCreator:
         phases_factory: Literal[False],
         phase_kws: Mapping[str, Any] | None = ...,
         segment_kws: Mapping[str, Any] | None = ...,
-        lnPiFreeEnergy_kws: Mapping[str, Any] | None = ...,
+        free_energy_kws: Mapping[str, Any] | None = ...,
         merge_kws: Mapping[str, Any] | None = ...,
         tag_phases: TagPhasesSignature | None = ...,
     ) -> tuple[list[lnPiMasked], MyNDArray]:
@@ -656,14 +651,14 @@ class PhaseCreator:
         phases_factory: PhasesFactorySignature | bool,
         phase_kws: Mapping[str, Any] | None = ...,
         segment_kws: Mapping[str, Any] | None = ...,
-        lnPiFreeEnergy_kws: Mapping[str, Any] | None = ...,
+        free_energy_kws: Mapping[str, Any] | None = ...,
         merge_kws: Mapping[str, Any] | None = ...,
         tag_phases: TagPhasesSignature | None = ...,
     ) -> tuple[list[lnPiMasked], MyNDArray] | lnPiCollection:
         ...
 
     @docfiller_local
-    def build_phases(
+    def build_phases(  # noqa: C901,PLR0912,PLR0913
         self,
         lnz: float | Sequence[float] | ArrayLike | MyNDArray | None = None,
         ref: lnPiMasked | None = None,
@@ -678,7 +673,7 @@ class PhaseCreator:
         phases_factory: PhasesFactorySignature | bool = True,
         phase_kws: Mapping[str, Any] | None = None,
         segment_kws: Mapping[str, Any] | None = None,
-        lnPiFreeEnergy_kws: Mapping[str, Any] | None = None,
+        free_energy_kws: Mapping[str, Any] | None = None,
         merge_kws: Mapping[str, Any] | None = None,
         tag_phases: TagPhasesSignature | None = None,
     ) -> tuple[list[lnPiMasked], MyNDArray] | lnPiCollection:
@@ -716,7 +711,7 @@ class PhaseCreator:
             Extra arguments to `phases_factory`
         segment_kws : mapping, optional
             Extra arguments to `self.segmenter.segment_lnpi`
-        lnPiFreeEnergy_kws : mapping, optional
+        free_energy_kws : mapping, optional
             Extra arguments to free energy calculation
         merge_kws : mapping, optional
             Extra arguments to merge
@@ -739,7 +734,8 @@ class PhaseCreator:
 
         if ref is None:
             if self.ref is None:
-                raise ValueError("must specify ref or self.ref")
+                msg = "must specify ref or self.ref"
+                raise ValueError(msg)
             ref = self.ref
 
         # reweight
@@ -769,11 +765,11 @@ class PhaseCreator:
             labels = self.segmenter.segment_lnpi(lnpi=ref, **segment_kws)
 
             # analyze w = - lnPi
-            lnPiFreeEnergy_kws = _combine_kws(
-                self.lnPiFreeEnergy_kws, lnPiFreeEnergy_kws, **connectivity_kws
+            free_energy_kws = _combine_kws(
+                self.free_energy_kws, free_energy_kws, **connectivity_kws
             )
             wlnpi = wFreeEnergy.from_labels(
-                data=ref.data, labels=labels, **lnPiFreeEnergy_kws
+                data=ref.data, labels=labels, **free_energy_kws
             )
 
             if merge_phases:
@@ -910,6 +906,10 @@ class BuildPhasesBase:
     def x(self) -> list[float | None]:
         return self._x
 
+    @x.setter
+    def x(self, x: list[float | None]) -> None:
+        self._set_x(x)
+
     @property
     def phase_creator(self) -> PhaseCreator:
         return self._phase_creator
@@ -921,10 +921,6 @@ class BuildPhasesBase:
         self._index = self._x.index(None)
         self._set_params()
 
-    @x.setter  # type: ignore
-    def x(self, x: list[float | None]) -> None:
-        self._set_x(x)
-
     @property
     def index(self) -> int:
         """Index number which varies"""
@@ -933,7 +929,7 @@ class BuildPhasesBase:
     def _set_params(self) -> None:
         pass
 
-    def _get_lnz(self, lnz_index: float) -> MyNDArray:  # pyright: ignore
+    def _get_lnz(self, lnz_index: float) -> MyNDArray:
         # to be implemented in child class
         raise NotImplementedError
 
@@ -1003,7 +999,7 @@ class BuildPhasesBase:
 
 # from .utils import get_lnz_iter
 @docfiller_local
-class BuildPhases_mu(BuildPhasesBase):
+class BuildPhases_mu(BuildPhasesBase):  # noqa: N801
     """
     create phases from scalar value of mu for fixed value of mu for other species
 
@@ -1023,7 +1019,7 @@ class BuildPhases_mu(BuildPhasesBase):
 
 
 @docfiller_local
-class BuildPhases_dmu(BuildPhasesBase):
+class BuildPhases_dmu(BuildPhasesBase):  # noqa: N801
     """
     Create phases from scalar value of mu at fixed value of dmu for other species
 
@@ -1044,5 +1040,5 @@ class BuildPhases_dmu(BuildPhasesBase):
 
 
 @lru_cache(maxsize=10)
-def get_default_PhaseCreator(nmax: int) -> PhaseCreator:
+def get_default_phasecreator(nmax: int) -> PhaseCreator:
     return PhaseCreator(nmax=nmax)
