@@ -6,6 +6,7 @@ Utility functions (:mod:`~lnpy.utils`)
 from __future__ import annotations
 
 from functools import lru_cache
+from itertools import starmap
 from typing import TYPE_CHECKING, TypedDict, cast, overload
 
 import numpy as np
@@ -19,7 +20,7 @@ if TYPE_CHECKING:
 
     import xarray as xr
     from numpy.typing import ArrayLike, DTypeLike, NDArray
-    from scipy.optimize import RootResults  # pyright: ignore
+    from scipy.optimize import RootResults
 
     from lnpy.lnpidata import lnPiMasked
 
@@ -42,20 +43,21 @@ def _get_tqdm_default() -> Callable[..., Any]:
     _tqdm = _get_tqdm()
     if _tqdm:
         try:
-            from IPython.core.getipython import get_ipython  # pyright: ignore
+            from IPython.core.getipython import (  # pyright: ignore[reportMissingImports]
+                get_ipython,
+            )
 
             p = get_ipython()  # type: ignore[no-untyped-call, unused-ignore]
             if p is not None and p.has_trait("kernel"):
                 from tqdm.notebook import tqdm as tqdm_default
 
-                return tqdm_default  # pyright: ignore
-            else:
-                return cast("Callable[..., Any]", _tqdm.tqdm)
+                return tqdm_default
+            return cast("Callable[..., Any]", _tqdm.tqdm)
         except ImportError:
             return cast("Callable[..., Any]", _tqdm.tqdm)
     else:
 
-        def wrapper(seq: Iterable[T], *args: Any, **kwargs: Any) -> Iterable[T]:
+        def wrapper(seq: Iterable[T], *args: Any, **kwargs: Any) -> Iterable[T]:  # noqa: ARG001
             return seq
 
         return wrapper
@@ -83,13 +85,13 @@ def get_tqdm(
     _tqdm = _get_tqdm()
 
     if isinstance(len_min, str):
-        len_min = OPTIONS[len_min]  # type: ignore
+        len_min = OPTIONS[len_min]  # type: ignore[literal-required]
 
     if n is None:
         seq = tuple(seq)
         n = len(seq)
 
-    if _tqdm and OPTIONS["tqdm_use"] and n >= len_min:  # pyright: ignore
+    if _tqdm and OPTIONS["tqdm_use"] and n >= len_min:
         if leave is None:
             leave = OPTIONS["tqdm_leave"]
         seq = tqdm(seq, leave=leave, **kwargs)
@@ -128,7 +130,7 @@ def get_tqdm_build(
 @lru_cache
 def _get_joblib() -> Any:
     try:
-        import joblib  # pyright: ignore
+        import joblib
     except ImportError:
         joblib = None
     return joblib
@@ -146,15 +148,14 @@ def _use_joblib(
         if total is None:
             total = len(items)
 
-        return total >= OPTIONS[len_key]  # type: ignore
-    else:
-        return False
+        return total >= OPTIONS[len_key]  # type: ignore[no-any-return,literal-required]
+    return False
 
 
 def _parallel(seq: Iterable[Any]) -> list[Any]:
     joblib = _get_joblib()
 
-    return joblib.Parallel(  # type: ignore
+    return joblib.Parallel(  # type: ignore[no-any-return]
         n_jobs=OPTIONS["joblib_n_jobs"],
         backend=OPTIONS["joblib_backend"],
         **OPTIONS["joblib_kws"],
@@ -170,8 +171,7 @@ def parallel_map_build(
 
     if _use_joblib(items, "joblib_len_build"):
         return _parallel(joblib.delayed(func)(x, *args, **kwargs) for x in items)
-    else:
-        return [func(x, *args, **kwargs) for x in items]
+    return [func(x, *args, **kwargs) for x in items]
 
 
 def _func_call(x: Callable[..., R], *args: Any, **kwargs: Any) -> R:
@@ -179,16 +179,18 @@ def _func_call(x: Callable[..., R], *args: Any, **kwargs: Any) -> R:
 
 
 def parallel_map_call(
-    items: Sequence[Callable[..., Any]], use_joblib: bool, *args: Any, **kwargs: Any
+    items: Sequence[Callable[..., Any]],
+    use_joblib: bool,  # noqa: ARG001
+    *args: Any,
+    **kwargs: Any,
 ) -> list[Any]:
     joblib = _get_joblib()
     if _use_joblib(items, "joblib_len_calc"):
         return _parallel(joblib.delayed(_func_call)(x, *args, **kwargs) for x in items)
-    else:
-        return [x(*args, **kwargs) for x in items]
+    return [x(*args, **kwargs) for x in items]
 
 
-def parallel_map_attr(attr: str, use_joblib: bool, items: Sequence[Any]) -> list[Any]:
+def parallel_map_attr(attr: str, use_joblib: bool, items: Sequence[Any]) -> list[Any]:  # noqa: ARG001
     from operator import attrgetter
 
     joblib = _get_joblib()
@@ -196,13 +198,12 @@ def parallel_map_attr(attr: str, use_joblib: bool, items: Sequence[Any]) -> list
 
     if _use_joblib(items, "joblib_len_calc"):
         return _parallel(joblib.delayed(func)(x) for x in items)
-    else:
-        return [func(x) for x in items]
+    return [func(x) for x in items]
 
 
 def parallel_map_func_starargs(
     func: Callable[..., R],
-    use_joblib: bool,
+    use_joblib: bool,  # noqa: ARG001
     items: Iterable[Any],
     total: int | None = None,
 ) -> list[R]:
@@ -211,9 +212,8 @@ def parallel_map_func_starargs(
     items = tuple(items)
 
     if _use_joblib(items, "joblib_len_calc", total=total):
-        return _parallel(joblib.delayed(func)(*x) for x in items)
-    else:
-        return [func(*x) for x in items]
+        return _parallel(starmap(joblib.delayed(func), items))
+    return list(starmap(func, items))
 
 
 # ----------------------------------------
@@ -229,11 +229,10 @@ def dim_to_suffix_dataarray(
     da: xr.DataArray, dim: Hashable, join: str = "_"
 ) -> xr.Dataset:
     if dim in da.dims:
-        return da.assign_coords(
-            **{dim: lambda x: [f"{x.name}{join}{c}" for c in x[dim].values]}  # type: ignore
+        return da.assign_coords(  # type: ignore[misc]
+            **{dim: lambda x: [f"{x.name}{join}{c}" for c in x[dim].values]}  # type: ignore[arg-type]
         ).to_dataset(dim=dim)
-    else:
-        return da.to_dataset()
+    return da.to_dataset()
 
 
 def dim_to_suffix_dataset(
@@ -242,9 +241,9 @@ def dim_to_suffix_dataset(
     out = table
     for k in out:
         if dim in out[k].dims:
-            out = out.drop_vars(k).update(
+            out = out.drop_vars(k).update(  # type: ignore[arg-type,unused-ignore]
                 table[k].pipe(dim_to_suffix_dataarray, dim, join)
-            )  # pyright: ignore
+            )
     return out
 
 
@@ -267,10 +266,10 @@ def dim_to_suffix(
 
     if isinstance(ds, DataArray):
         return dim_to_suffix_dataarray(ds, dim=dim, join=join)
-    elif isinstance(ds, Dataset):  # pyright: ignore
+    if isinstance(ds, Dataset):
         return dim_to_suffix_dataset(ds, dim=dim, join=join)
-    else:
-        raise ValueError("`ds` must be `DataArray` or `Dataset`")
+    msg = "`ds` must be `DataArray` or `Dataset`"
+    raise ValueError(msg)
 
 
 def _convention_to_bool(convention: MaskConvention) -> bool:
@@ -278,8 +277,9 @@ def _convention_to_bool(convention: MaskConvention) -> bool:
         convention = True
     elif convention == "masked":
         convention = False
-    elif not isinstance(convention, bool):  # pyright: ignore
-        raise ValueError(f"Bad value {convention} sent to _convention_to_bool")
+    elif not isinstance(convention, bool):
+        msg = f"Bad value {convention} sent to _convention_to_bool"
+        raise ValueError(msg)
     return convention
 
 
@@ -459,7 +459,7 @@ def labels_to_masks(
     from skimage import segmentation
 
     if include_boundary:
-        kwargs = dict(dict(mode="outer", connectivity=labels.ndim), **kwargs)
+        kwargs = dict({"mode": "outer", "connectivity": labels.ndim}, **kwargs)
     if features is None:
         features = [i for i in np.unique(labels) if i > 0]
 
@@ -473,15 +473,17 @@ def labels_to_masks(
     for i in features:
         m: NDArray[np.bool_] = labels == i
         if include_boundary:
+            # fmt: off
             b = cast(
                 "NDArray[np.bool_]",
-                segmentation.find_boundaries(m.astype(int), **kwargs),
-            )  # pyright: ignore[reportUnknownMemberType]
+                segmentation.find_boundaries(m.astype(int), **kwargs),  # pyright: ignore[reportUnknownMemberType]
+            )
+            # fmt: on
             m = m | b
         if not convention:
             m = ~m
         output.append(m)
-    return output, features  # type: ignore
+    return output, features  # type: ignore[return-value]
 
 
 @docfiller.decorate
@@ -490,7 +492,6 @@ def masks_to_labels(
     features: Sequence[int] | MyNDArray | None = None,
     convention: MaskConvention = "image",
     dtype: DTypeLike = np.int_,
-    **kwargs: Any,
 ) -> MyNDArray:
     """
     Convert list of masks to labels
@@ -528,24 +529,24 @@ def masks_to_labels(
 
 
 def ffill(arr: MyNDArray, axis: int = -1, limit: int | None = None) -> MyNDArray:
-    import bottleneck  # pyright: ignore
+    import bottleneck
 
     _limit = limit if limit is not None else arr.shape[axis]
-    return bottleneck.push(arr, n=_limit, axis=axis)  # type: ignore
+    return bottleneck.push(arr, n=_limit, axis=axis)  # type: ignore[no-any-return]
 
 
 def bfill(arr: MyNDArray, axis: int = -1, limit: int | None = None) -> MyNDArray:
     """Inverse of ffill"""
-    import bottleneck  # pyright: ignore
+    import bottleneck
 
     # work around for bottleneck 178
     _limit = limit if limit is not None else arr.shape[axis]
 
     arr = np.flip(arr, axis=axis)
     # fill
-    arr = bottleneck.push(arr, axis=axis, n=_limit)  # pyright: ignore
+    arr = bottleneck.push(arr, axis=axis, n=_limit)
     # reverse back to original
-    return np.flip(arr, axis=axis)  # pyright: ignore
+    return np.flip(arr, axis=axis)
 
 
 ##################################################
@@ -576,14 +577,7 @@ def get_lnz_iter(lnz: Iterable[float | None], x: ArrayLike) -> MyNDArray:
     x = np.asarray(x)
     z = np.zeros_like(x)
 
-    L: list[MyNDArray] = []
-    for m in lnz:
-        if m is None:
-            L.append(x)
-        else:
-            L.append(z + m)
-
-    return np.array(L).T
+    return np.array([x if m is None else z + m for m in lnz]).T
 
 
 ##################################################
@@ -608,7 +602,7 @@ def get_lnz_iter(lnz: Iterable[float | None], x: ArrayLike) -> MyNDArray:
 #         Objects in sorted order.
 #     """
 
-#     molfrac_comp = np.array([x.molfrac[comp] for x in values]) # type: ignore
+#     molfrac_comp = np.array([x.molfrac[comp] for x in values])
 
 #     order = np.argsort(molfrac_comp)
 
@@ -642,7 +636,7 @@ def distance_matrix(
     --------
     ~scipy.ndimage.distance_transform_edt
     """
-    from scipy.ndimage import distance_transform_edt  # pyright: ignore
+    from scipy.ndimage import distance_transform_edt
 
     mask = np.asarray(mask, dtype=bool)
     mask = mask_change_convention(mask, convention_in=convention, convention_out=True)
@@ -654,11 +648,11 @@ def distance_matrix(
     mask = np.pad(mask, pad_width=pad_width, mode="constant", constant_values=False)
 
     # distance filter
-    dist = distance_transform_edt(mask)  # pyright: ignore
+    dist = distance_transform_edt(mask)
 
     # remove padding
     s = (slice(None, -1),) * ndim
-    return dist[s]  # type: ignore
+    return dist[s]  # type: ignore[no-any-return]
 
 
 def lnpimasked_to_dataset(
@@ -710,7 +704,7 @@ def dataset_to_lnpimasked(
         extra_kws = {}
 
     if pe_name in ds:
-        extra_kws[pe_name] = ds[pe_name].values  # type: ignore
+        extra_kws[pe_name] = ds[pe_name].values  # type: ignore[index]
 
     return lnPiMasked.from_dataarray(da=data, extra_kws=extra_kws, **kwargs)
 
@@ -741,7 +735,7 @@ def rootresults_to_rootresultdict(
         root=r.root,
         iterations=r.iterations,
         function_calls=r.function_calls,
-        converged=r.converged,  # pyright: ignore
+        converged=r.converged,
         flag=r.flag,
     )
 
