@@ -3,21 +3,17 @@ from __future__ import annotations
 
 
 def get_conda_environment_map(simplify: bool = True) -> dict[str, str]:
-    """Contruct mapping from environment env_name to path"""
+    """Construct mapping from environment env_name to path"""
     import subprocess
     from pathlib import Path
 
     result = subprocess.check_output(["conda", "env", "list"])
 
-    env_map = {}
+    env_map: dict[str, str] = {}
     for line in result.decode().split("\n"):
         if not line.startswith("#"):
             x = line.replace("*", "").split()
-            n = len(x)
-            if n == 1:
-                # this is path environment
-                pass
-            elif n == 2:
+            if len(x) == 2:
                 env_map[x[0]] = x[1]
     if simplify:
         home = str(Path.home())
