@@ -93,12 +93,30 @@ class ProjectConfig:
         if path.exists():
             python_paths, env_extras = cls._path_to_params(path)
             return cls(python_paths=python_paths, env_extras=env_extras)
-        else:
-            return cls()
+
+        return cls()
+
+    @classmethod
+    def from_path_and_environ(
+        cls,
+        path: str | Path = "./config/userconfig.toml",
+    ) -> Self:
+        def _get_python_paths_from_environ() -> list[str]:
+            if python_paths_environ := os.environ.get("NOX_PYTHON_PATH"):
+                return python_paths_environ.split(":")
+            return []
+
+        python_paths, env_extras = cls._path_to_params(path)
+
+        return cls(
+            python_paths=(python_paths or _get_python_paths_from_environ()),
+            env_extras=env_extras,
+        )
 
     @staticmethod
     def _params_to_string(
-        python_paths: list[str], env_extras: dict[str, Mapping[str, Any]]
+        python_paths: list[str],
+        env_extras: dict[str, Mapping[str, Any]],
     ) -> str:
         import configparser
         import json
@@ -133,20 +151,21 @@ class ProjectConfig:
         # [nox.python]
         # paths = ["~/.conda/envs/python-3.*/bin"]
         #
-        # [tool.pyproject2conda.envs.dev]
+        # [tool.pyproject2conda.envs.dev-user]
         # extras = ["dev-complete"]
-        """
+        """,
         )
 
         return header + s
 
     def to_path(self, path: str | Path | None = None) -> str:
         s = self._params_to_string(
-            python_paths=self.python_paths, env_extras=self.env_extras
+            python_paths=self.python_paths,
+            env_extras=self.env_extras,
         )
 
         if path is not None:
-            with open(path, "w") as f:
+            with Path(path).open("w") as f:
                 f.write(s)
         return s
 
@@ -156,25 +175,26 @@ class ProjectConfig:
     def expand_python_paths(self) -> list[str]:
         from glob import glob
 
-        paths = []
+        paths: list[str] = []
         for p in self.python_paths:
-            paths.extend(glob(os.path.expanduser(p)))
+            paths.extend(glob(os.path.expanduser(p)))  # noqa: PTH207, PTH111
         return paths
 
     def add_paths_to_environ(
-        self, paths: list[str] | None, prepend: bool = True
+        self,
+        paths: list[str] | None,
+        prepend: bool = True,
     ) -> None:
         if paths is None:
             paths = self.expand_python_paths()
         paths_str = ":".join(map(str, paths))
-        if prepend:
-            fmt = "{path_new}:{path_old}"
-        else:
-            fmt = "{path_old}:{path_new}"
+        fmt = "{path_new}:{path_old}" if prepend else "{path_old}:{path_new}"
         os.environ["PATH"] = fmt.format(path_new=paths_str, path_old=os.environ["PATH"])
 
     def to_nox_config(
-        self, add_paths_to_environ: bool = True, prepend: bool = True
+        self,
+        add_paths_to_environ: bool = True,
+        prepend: bool = True,
     ) -> dict[str, Any]:
         config: dict[str, Any] = {}
 
@@ -187,7 +207,7 @@ class ProjectConfig:
         if self.env_extras:
             config["environment-extras"] = self.env_extras
         else:
-            config["environment-extras"] = {"dev": ["nox", "dev"]}
+            config["environment-extras"] = {"dev-user": ["nox", "dev"]}
 
         return config
 
@@ -199,7 +219,7 @@ def glob_envs_to_paths(globs: list[str]) -> list[str]:
 
     env_map = get_conda_environment_map()
 
-    out = []
+    out: list[str] = []
     for glob in globs:
         found_envs = fnmatch.filter(env_map.keys(), glob)
         out.extend([f"{env_map[k]}/bin" for k in found_envs])
@@ -233,7 +253,7 @@ def main() -> None:
         "-d",
         "--dev-extras",
         nargs="+",
-        help="extras (from pyproject.toml) to include in development environment",
+        help="extras (from pyproject.toml) to include in `dev-user` environment",
     )
 
     p.add_argument(
@@ -259,18 +279,18 @@ def main() -> None:
         n.python_paths = python_paths
 
     if args.dev_extras:
-        n.env_extras["tool.pyproject2conda.envs.dev"] = {"extras": args.dev_extras}
+        n.env_extras["tool.pyproject2conda.envs.dev-user"] = {"extras": args.dev_extras}
 
     n.to_path(args.file)
 
 
 if __name__ == "__main__":
-    if __package__ is None:
+    if __package__ is None:  # pyright: ignore[reportUnnecessaryComparison]
         # Magic to be able to run script as either
         #   $ python -m tools.create_python
         # or
         #   $ python tools/create_python.py
         here = Path(__file__).absolute()
-        __package__ = here.parent.name
+        __package__ = here.parent.name  # noqa: A001
 
     main()
